feat: update navbar UI per Figma
