+++ packages/app/src/navigation/SidebarNavigationRow.vue
@@ -1,44 +1,31 @@
 <template>
   <SidebarTooltip
     :class="active
-      ? 'before:bg-jade-300 before:scale-x-100 before:transition-colors'
-      : 'before:scale-x-0 before:transition-transform'"
-    class="w-full
-      min-w-40px
-      relative
+      ? 'before:(bg-indigo-300 scale-x-100 transition-colors) cursor-default'
+      : 'before:(scale-x-0 transition-transform bg-gray-300)'"
+    class="rounded-md
       flex
-      items-center
-      rounded-md
-      group
       h-40px
       my-16px
+      w-full
+      min-w-40px
+      relative
+      items-center
+      group
       focus-visible:outline-none
-      before:content-open-square
-      before:mr-4px
-      before:rounded-r-md
-      before:text-transparent
-      before:h-40px
-      before:w-4px
-      before:bg-indigo-300
-      before:transform before:origin-left
-      before:duration-300
-      hover:before:scale-x-100
-      hover:before:bg-indigo-300
-      focus:before:bg-indigo-300"
+      before:(rounded-r-md h-40px mr-4px text-transparent transform origin-left w-4px duration-300 content-open-square) hover:before:scale-x-100 "
     :disabled="mainStore.navBarExpanded"
   >
     <component
       :is="
         icon"
-      :class="active ? 'icon-dark-jade-300 icon-light-jade-800' : 'icon-dark-gray-500 icon-light-gray-900'"
-      class="w-24px h-24px flex-shrink-0 m-12px
-            group-hover:icon-dark-indigo-300 group-hover:icon-light-indigo-600
-            group-focus:icon-dark-indigo-300 group-focus:icon-light-indigo-600
+      :class="active ? 'icon-dark-indigo-300 icon-light-indigo-700' : 'icon-dark-gray-500 icon-light-gray-900 group-hover:(icon-dark-gray-300 icon-light-gray-800) group-focus:(icon-dark-gray-300 icon-light-gray-800)'"
+      class="flex-shrink-0 h-24px m-12px w-24px
             children:transition children:duration-300"
     />
     <span
-      :class="[active ? 'text-jade-300' : 'text-gray-500']"
-      class="ml-8px truncate group-hocus:text-indigo-300 transition-colors duration-300"
+      :class="[active ? 'text-indigo-300' : 'text-gray-500 group-hocus:text-gray-300']"
+      class="ml-8px transition-colors duration-300 truncate"
     >
       {{ name }}
     </span>
