+++ packages/frontend-shared/cypress/e2e/support/e2eProjectDirs.ts
@@ -31,6 +31,9 @@ export const e2eProjectDirs = [
   'max-listeners',
   'migration',
   'migration-component-testing',
+  'migration-e2e-custom-integration',
+  'migration-e2e-custom-test-files',
+  'migration-e2e-fully-custom',
   'multiple-config-files-with-json',
   'multiple-support-files',
   'multiple-task-registrations',
